refactor(core): move Coach and CricketCoach to util package and update component scanning to include all the base packages
