+++ 02-spring-boot-spring-core/02-component-scanning/src/main/java/np/com/krishnabk/springcoredemo/Application.java
@@ -3,7 +3,10 @@ package np.com.krishnabk.springcoredemo;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-@SpringBootApplication
+@SpringBootApplication(
+    scanBasePackages = {"np.com.krishnabk.springcoredemo",
+                        "np.com.krishnabk.util"
+    })
 public class Application {
 
     public static void main(String[] args) {
