+++ 02-spring-boot-spring-core/02-component-scanning/src/main/java/np/com/krishnabk/util/CricketCoach.java
@@ -1,4 +1,4 @@
-package np.com.krishnabk.springcoredemo.common;
+package np.com.krishnabk.util;
 
 import org.springframework.stereotype.Component;
